# Conflicts:
#   backend/src/main/java/com/ssafy/gaese/domain/user/entity/Ability.java

+++ backend/src/main/java/com/ssafy/gaese/domain/algorithm/dto/AlgoSolveReq.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class AlgoSolveReq {
-    int problemId;
+    Long problemId;
     String userBjId;
     int lanId;
 }
